REST endpoints of sessions
